+++ onsiteprint-plugin/blocks/event/block-script-parts/setup-header.js
@@ -1,8 +1,8 @@
 /* ------------------------------------------------------------------------
  #  JS Part Name: Setup Header
  *  Block function included in the Event Block.
- ?  Updated: 2025-07-03 - 01:18 (Y:m:d - H:i)
- ?  Info: Added new Setup Header Script with Function.
+ ?  Updated: 2025-07-28 - 01:57 (Y:m:d - H:i)
+ ?  Info: Added opModalToggleListener() and opCreateParticipantListener().
 
 ---------------------------------------------------------------------------
  #  TABLE OF CONTENTS:
@@ -16,27 +16,44 @@
  #  1. Import Functions from Scripts
 --------------------------------------------------------------------------- */
 import * as opModuleBasic from '../../../assets/js/inc/basic.js'
+import { opModalToggleListener } from '../../../assets/js/inc/modal/toggle-modal-listener.js'
+import { opCreateParticipantListener } from './participant-listeners.js'
 
 /* ------------------------------------------------------------------------
  #  2. Function: Setup the Header of the Event
  ?  NB: The function is under construction.
 --------------------------------------------------------------------------- */
-export function opSetupHeader( debug, block ) {
+export function opSetupHeader( debug, block, eventId, columnAmount ) {
 
     try {
         
         ///// Get Function Name.
         var functionName = opSetupHeader.name
-        
+
         ///// Set the Debug.
         ////* Set the Parameter If is not defined (true or false).
         if ( debug !== true ) debug = false
         if ( debug ) console.group( `${ functionName }()` )
-   
-            console.log('Header Works!!')
 
-        ///// End the Console Log Group.
-        if ( debug ) console.groupEnd()
+        ///// Get the Create Participant Template Element.
+        let modalTemplateElement = block.querySelector( `[id$="-create-participant-template"]` )
+        let modalElement = modalTemplateElement.content.cloneNode(true)
+        let modalHeader = modalElement.querySelector( '.op-header-content__inner' )
+        let modalContent = modalElement.querySelector( '.op-modal-content__inner' )
+        let formElement = modalContent.querySelector( '.op-form' )
+        let columnInputElements = formElement.querySelectorAll( '.op-col-input input' )
+
+        ///// Set Modal Toggle Listener to the Add Participant Button and Modal Close Button.
+        opModalToggleListener( debug, block.querySelector( '.op-button-cancel' ), false )
+        opModalToggleListener( debug, block.querySelector( '.op-button-add' ), true, modalHeader, modalContent )
+
+        ///// Set Create Participant Listener to the Modal Form Button.
+        opCreateParticipantListener( debug, block, modalContent.querySelector( '.op-button-save' ), eventId, formElement )
+
+        /* ///// Set Column Input Listener to All of the Column Input Elements in the Modal Form.
+        columnInputElements.forEach( inputElement => {
+            opColumnInputListener( debug, inputElement )
+        }) */
 
         ///// Return the Response.
         return opModuleBasic.opReturnResponse( false, 200, {